+++ lib/consumer.dart
@@ -12,31 +12,27 @@ import 'package:flutter/material.dart';
 /// setState is trigger all stream listen, update state and call all Consumer widget update.
 /// build is use _ConsumerWidget create a StatefulWidget, and subscribe consumer at _ConsumerWidget.
 class Consumer<T> {
-  T state;
-  StreamController controller;
+  final T state;
+  StreamController _controller;
   Stream stream;
 
   Consumer(this.state) {
-    controller = StreamController.broadcast();
-    stream = controller.stream;
-    stream.listen((data) {
-      state = data;
-    });
+    _controller = StreamController.broadcast();
+    stream = _controller.stream;
   }
 
-  Widget build(Widget Function(BuildContext ctx, T state) builder, {List<dynamic> Function(T s) memo}) {
+  Widget build(Widget Function(BuildContext ctx, T state) builder,
+      {@required List<dynamic> Function(T s) memo}) {
     return _ConsumerWidget<T>(ctrl: this, memo: memo, builder: builder);
   }
 
-  @Deprecated('remove getState to v2.2.0, use consumer.state replace consumer.getState()')
-  T getState() {
-    return state;
-  }
-
-  T setState(Function(T state) fn) {
+  // @Deprecated('remove getState to v2.2.0, use consumer.state replace consumer.getState()')
+  // T getState() {
+  //   return state;
+  // }
+  setState(Function(T state) fn) {
     fn(state);
-    controller.add(state);
-    return state;
+    _controller.add(state);
   }
 }
 
@@ -52,10 +48,16 @@ class _ConsumerWidget<T> extends StatefulWidget {
   final List<dynamic> Function(T state) memo;
   final Widget Function(BuildContext ctx, T state) builder;
 
-  _ConsumerWidget({@required this.ctrl, @required this.builder, @required this.memo, Key key}) : super(key: key);
+  _ConsumerWidget(
+      {@required this.ctrl,
+      @required this.builder,
+      @required this.memo,
+      Key key})
+      : super(key: key);
 
   @override
-  _ConsumerWidgetState createState() => _ConsumerWidgetState<T>(ctrl, memo, builder);
+  _ConsumerWidgetState createState() =>
+      _ConsumerWidgetState<T>(ctrl, memo, builder);
 }
 
 class _ConsumerWidgetState<T> extends State<_ConsumerWidget> {
@@ -70,15 +72,9 @@ class _ConsumerWidgetState<T> extends State<_ConsumerWidget> {
   @override
   void initState() {
     super.initState();
-    if (_memo != null) {
-      _lastMemo = [..._memo(_ctrl.state)];
-    }
+    _lastMemo = [..._memo(_ctrl.state)];
 
     _sub = _ctrl.stream.listen((data) {
-      if (_memo == null) {
-        setState(() {});
-        return;
-      }
       if (_lastMemo.length > 0) {
         bool isUpdate = false;
         List nowMemo = [..._memo(_ctrl.state)];